Add cmdletbinding to all functions.
The default format for any function should be to support the
common parameters and other functionality added by cmdletbinding.

Discussions and information can be found on the community best practice
repository:
https://github.com/PoshCode/PowerShellPracticeAndStyle/blob/master/Style-Guide/Function-Structure.md
with disucssion here:
https://github.com/PoshCode/PowerShellPracticeAndStyle/issues/32

general consensus was that cmdletbinding should be the default approach
with "basic" functions chosen when they were specifically required.